pyproject.toml
[project]
name = "rag_llm"
version = "0.1.0"
description = "RAG-based Private Knowledge Worker"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.3",
    "faiss-cpu>=1.7.4",
    "PyPDF2>=3.0.1",
    "python-docx>=0.8.11",
    "beautifulsoup4>=4.12.2",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.7.0"]
huggingface = ["transformers>=4.30.0"]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
rag-llm = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
